Cleared all test_mains and transferred important codes to code_snippets.cpp
